+++ tests/test_gui_structure_elements/test_aim.py
@@ -26,5 +26,9 @@ def test_aim(qtbot):
 
     assert main_window.gui_structure.aim_sub.widget.isChecked()
 
+    main_window.save_scenario()
+    main_window.gui_structure.aim_sub.widget.click()
+    assert main_window.list_widget_scenario.currentItem().text()[-1] == "*"
+
     main_window.gui_structure.aim_plot.set_text("Hello")
     assert main_window.gui_structure.aim_plot.widget.text() == "Hello"