+++ tests/test_gui_structure_elements/test_results_figure.py
@@ -2,11 +2,14 @@ from __future__ import annotations
 
 import os
 from pathlib import Path
+from platform import system
 
 import numpy as np
 from matplotlib.backends import qt_compat
 
+from ScenarioGUI.gui_classes.gui_structure_classes.result_figure import font_list_by_name
 from tests.starting_closing_tests import close_tests, start_tests
+import ScenarioGUI.global_settings as globs
 
 
 class Event:
@@ -95,20 +98,22 @@ def test_results_figure(qtbot):  # noqa: PLR0915
     main_window.display_results()
     main_window.gui_structure.figure_results_with_customizable_layout.change_font()
 
+    font_idx = font_list_by_name.index(("Verdana" if system() == "Windows" else "Arial").upper())
+
     main_window.gui_structure.figure_results_with_customizable_layout.option_figure_background.set_value((100, 110, 111))
     main_window.gui_structure.figure_results_with_customizable_layout.option_plot_background.set_value((101, 111, 112))
     main_window.gui_structure.figure_results_with_customizable_layout.option_axes_text.set_value((99, 113, 115))
     main_window.gui_structure.figure_results_with_customizable_layout.option_axes.set_value((89, 90, 91))
-    main_window.gui_structure.figure_results_with_customizable_layout.option_font.set_value(3)
-    main_window.gui_structure.figure_results_with_customizable_layout.option_font_size.set_value(15)
+    main_window.gui_structure.figure_results_with_customizable_layout.option_font.set_value(font_idx)
+    main_window.gui_structure.figure_results_with_customizable_layout.option_font_size.set_value(10)
     main_window.gui_structure.figure_results_with_customizable_layout.option_legend_text.set_value((120, 121, 122))
     main_window.gui_structure.figure_results_with_customizable_layout.option_title.set_value((130, 131, 132))
     assert not np.allclose(main_window.gui_structure.option_figure_background.get_value(), (100, 110, 111))
     assert not np.allclose(main_window.gui_structure.option_plot_background.get_value(), (100, 110, 111))
     assert not np.allclose(main_window.gui_structure.option_axes_text.get_value(), (100, 110, 111))
     assert not np.allclose(main_window.gui_structure.option_axes.get_value(), (100, 110, 111))
-    assert not np.isclose(main_window.gui_structure.option_font.get_value()[0], 3)
-    assert not np.isclose(main_window.gui_structure.option_font_size_figure.get_value(), 15)
+    assert not np.isclose(main_window.gui_structure.option_font.get_value()[0], font_idx)
+    assert not np.isclose(main_window.gui_structure.option_font_size_figure.get_value(), 10)
     assert not np.allclose(main_window.gui_structure.option_legend_text.get_value(), (120, 121, 122))
     assert not np.allclose(main_window.gui_structure.option_title.get_value(), (130, 131, 132))
     main_window.gui_structure.figure_results_with_customizable_layout.option_save_layout.button.click()
@@ -116,8 +121,8 @@ def test_results_figure(qtbot):  # noqa: PLR0915
     assert np.allclose(main_window.gui_structure.option_plot_background.get_value(), (101, 111, 112))
     assert np.allclose(main_window.gui_structure.option_axes_text.get_value(), (99, 113, 115))
     assert np.allclose(main_window.gui_structure.option_axes.get_value(), (89, 90, 91))
-    assert np.allclose(main_window.gui_structure.option_font.get_value()[0], 3)
-    assert np.isclose(main_window.gui_structure.option_font_size_figure.get_value(), 15)
+    assert np.allclose(main_window.gui_structure.option_font.get_value()[0], font_idx)
+    assert np.isclose(main_window.gui_structure.option_font_size_figure.get_value(), 10)
     assert np.allclose(main_window.gui_structure.option_legend_text.get_value(), (120, 121, 122))
     assert np.allclose(main_window.gui_structure.option_title.get_value(), (130, 131, 132))
 
@@ -125,8 +130,8 @@ def test_results_figure(qtbot):  # noqa: PLR0915
     assert np.allclose(main_window.gui_structure.figure_results_with_customizable_layout.option_plot_background.get_value(), (101, 111, 112))
     assert np.allclose(main_window.gui_structure.figure_results_with_customizable_layout.option_axes_text.get_value(), (99, 113, 115))
     assert np.allclose(main_window.gui_structure.figure_results_with_customizable_layout.option_axes.get_value(), (89, 90, 91))
-    assert np.allclose(main_window.gui_structure.figure_results_with_customizable_layout.option_font.get_value()[0], 3)
-    assert np.isclose(main_window.gui_structure.figure_results_with_customizable_layout.option_font_size.get_value(), 15)
+    assert np.allclose(main_window.gui_structure.figure_results_with_customizable_layout.option_font.get_value()[0], font_idx)
+    assert np.isclose(main_window.gui_structure.figure_results_with_customizable_layout.option_font_size.get_value(), 10)
     assert np.allclose(main_window.gui_structure.figure_results_with_customizable_layout.option_legend_text.get_value(), (120, 121, 122))
     assert np.allclose(main_window.gui_structure.figure_results_with_customizable_layout.option_title.get_value(), (130, 131, 132))
 
@@ -134,7 +139,7 @@ def test_results_figure(qtbot):  # noqa: PLR0915
     main_window.gui_structure.option_plot_background.set_value((13, 14, 15))
     main_window.gui_structure.option_axes_text.set_value((10, 11, 12))
     main_window.gui_structure.option_axes.set_value((7, 8, 9))
-    main_window.gui_structure.option_font.set_value(2)
+    main_window.gui_structure.option_font.set_value(font_list_by_name.index(globs.FONT.upper()))
     main_window.gui_structure.option_font_size_figure.set_value(14)
     main_window.gui_structure.option_legend_text.set_value((1, 2, 3))
     main_window.gui_structure.option_title.set_value((4, 5, 6))
@@ -143,7 +148,7 @@ def test_results_figure(qtbot):  # noqa: PLR0915
     assert np.allclose(main_window.gui_structure.figure_results_with_customizable_layout.option_plot_background.get_value(), (13, 14, 15))
     assert np.allclose(main_window.gui_structure.figure_results_with_customizable_layout.option_axes_text.get_value(), (10, 11, 12))
     assert np.allclose(main_window.gui_structure.figure_results_with_customizable_layout.option_axes.get_value(), (7, 8, 9))
-    assert np.allclose(main_window.gui_structure.figure_results_with_customizable_layout.option_font.get_value()[0], 2)
+    assert np.allclose(main_window.gui_structure.figure_results_with_customizable_layout.option_font.get_value()[0], font_list_by_name.index(globs.FONT.upper()))
     assert np.isclose(main_window.gui_structure.figure_results_with_customizable_layout.option_font_size.get_value(), 14)
     assert np.allclose(main_window.gui_structure.figure_results_with_customizable_layout.option_legend_text.get_value(), (1, 2, 3))
     assert np.allclose(main_window.gui_structure.figure_results_with_customizable_layout.option_title.get_value(), (4, 5, 6))